Add OpenTracing implementation using basictracer-go
This implementation uses opentracing/basictracer-go for the underlying
opentracing implementation. This patch adds a Recorder that writes spans
to an appdash.Collector, which can then be plugged into the
basictracer.Tracer.

The main benefit of using basictracer-go as the underlying
implementation is that we essentially remove all the redundant code use in
the opentracing1.0-rc1 branch.